sdlc(impl): Split agent.ts into stream.ts and claude-process.ts (FR-E30)
- Files changed: engine/agent.ts reduced from 798→297 lines by extracting
  stream processing (stream.ts, 216 lines) and CLI invocation
  (claude-process.ts, 241 lines) into dedicated modules
- engine/stream.ts: FileReadTracker, StreamProcessorState, processStreamEvent,
  extractClaudeOutput, formatEventForOutput, formatFooter, tsPrefix, stampLines
- engine/claude-process.ts: InvokeOptions, invokeClaudeCli, buildClaudeArgs,
  executeClaudeProcess
- engine/agent_test.ts: updated imports to new modules
- engine/hitl.ts: updated InvokeOptions import + dynamic import to claude-process.ts
- engine/mod.ts: updated InvokeOptions re-export to claude-process.ts
- Tests: no changes to test logic, all 478 pass
- Check: PASS
