+++ engine/agent.ts
@@ -15,7 +15,7 @@ import {
   type ValidationResult,
 } from "./validate.ts";
 import type { OutputManager, VerboseInput } from "./output.ts";
-import { register, unregister } from "./process-registry.ts";
+import { invokeClaudeCli } from "./claude-process.ts";
 
 /**
  * Resolve input artifact file paths and sizes from input directories.
@@ -266,470 +266,6 @@ export async function runAgent(opts: AgentRunOptions): Promise<AgentResult> {
 
 // --- Internal helpers ---
 
-/** Low-level options for a single claude CLI invocation (initial or resume). */
-export interface InvokeOptions {
-  /** Path to system prompt file appended via --append-system-prompt-file. */
-  promptFile?: string;
-  /** Cached prompt content (takes priority over promptFile). */
-  promptContent?: string;
-  /** Task prompt passed to claude via -p flag. */
-  taskPrompt: string;
-  /** Session ID for --resume continuation (omit for initial invocation). */
-  resumeSessionId?: string;
-  /** Extra CLI arguments passed to claude command. */
-  claudeArgs?: string[];
-  /** Claude model override. Skipped on resume (session inherits model). */
-  model?: string;
-  /** Max seconds before SIGTERM kills the claude process. */
-  timeoutSeconds: number;
-  /** Max retry attempts on CLI crash/error before giving up. */
-  maxRetries: number;
-  /** Base delay between retries in seconds (doubled each attempt). */
-  retryDelaySeconds: number;
-  /** Callback invoked with each formatted stream event line for terminal display. */
-  onOutput?: (line: string) => void;
-  /** Path to write real-time stream-json log file. */
-  streamLogPath?: string;
-  /** Verbosity level for terminal output filtering (semi-verbose suppresses tool_use). */
-  verbosity?: Verbosity;
-}
-
-interface InvokeResult {
-  output?: ClaudeCliOutput;
-  error?: string;
-}
-
-/** Invoke claude CLI with retry logic. */
-export async function invokeClaudeCli(
-  opts: InvokeOptions,
-): Promise<InvokeResult> {
-  const args = buildClaudeArgs(opts);
-  let lastError = "";
-
-  for (let attempt = 1; attempt <= opts.maxRetries; attempt++) {
-    try {
-      const output = await executeClaudeProcess(
-        args,
-        opts.timeoutSeconds,
-        opts.onOutput,
-        opts.streamLogPath,
-        opts.verbosity,
-      );
-      if (output.is_error) {
-        lastError = `Claude CLI returned error: ${output.result}`;
-        if (attempt < opts.maxRetries) {
-          const delay = opts.retryDelaySeconds * Math.pow(2, attempt - 1);
-          await sleep(delay * 1000);
-          continue;
-        }
-        return { output, error: lastError };
-      }
-      return { output };
-    } catch (err) {
-      lastError = (err as Error).message;
-      if (attempt < opts.maxRetries) {
-        const delay = opts.retryDelaySeconds * Math.pow(2, attempt - 1);
-        await sleep(delay * 1000);
-        continue;
-      }
-    }
-  }
-
-  return {
-    error: `Claude CLI failed after ${opts.maxRetries} attempts: ${lastError}`,
-  };
-}
-
-/** Build CLI arguments for the claude command. Exported for testing. */
-export function buildClaudeArgs(opts: InvokeOptions): string[] {
-  const args: string[] = [];
-
-  // Extra CLI args (e.g. --dangerously-skip-permissions) go first
-  if (opts.claudeArgs && opts.claudeArgs.length > 0) {
-    args.push(...opts.claudeArgs);
-  }
-
-  if (opts.resumeSessionId) {
-    args.push("--resume", opts.resumeSessionId);
-  }
-
-  args.push("-p", opts.taskPrompt);
-
-  if (!opts.resumeSessionId) {
-    if (opts.promptContent) {
-      args.push("--append-system-prompt", opts.promptContent);
-    } else if (opts.promptFile) {
-      args.push("--append-system-prompt-file", opts.promptFile);
-    }
-  }
-
-  if (opts.model && !opts.resumeSessionId) {
-    args.push("--model", opts.model);
-  }
-
-  args.push("--output-format", "stream-json", "--verbose");
-
-  return args;
-}
-
-/**
- * Tracks per-path file read counts within a single agent invocation.
- * Returns a warning string when a path is read more than `threshold` times.
- * Pure-logic class — unit-testable without I/O.
- */
-export class FileReadTracker {
-  private counts = new Map<string, number>();
-
-  constructor(private readonly threshold = 2) {}
-
-  /**
-   * Increment read count for path.
-   * Returns `[WARN] repeated file read: <path> (<N> times)` when count > threshold, else null.
-   */
-  track(path: string): string | null {
-    const count = (this.counts.get(path) ?? 0) + 1;
-    this.counts.set(path, count);
-    if (count > this.threshold) {
-      return `[WARN] repeated file read: ${path} (${count} times)`;
-    }
-    return null;
-  }
-
-  /** Clear all counts (for testing isolation). */
-  reset(): void {
-    this.counts.clear();
-  }
-}
-
-/** Mutable state bag for processStreamEvent() — holds all stream-processing state. */
-export interface StreamProcessorState {
-  /** Count of assistant turns seen so far (increments on each assistant event). */
-  turnCount: number;
-  /** Extracted result event; populated when a "result" event is processed. */
-  resultEvent: ClaudeCliOutput | undefined;
-  /** Tracks per-path file read counts to detect repeated reads. */
-  tracker: FileReadTracker;
-  /** Open log file handle for writing formatted summaries (undefined = no log). */
-  logFile: Deno.FsFile | undefined;
-  /** Text encoder shared across writes. */
-  encoder: TextEncoder;
-  /** Callback for forwarding verbosity-filtered event summaries to terminal. */
-  onOutput?: (line: string) => void;
-  /** Verbosity level controls which event types reach terminal output. */
-  verbosity?: Verbosity;
-}
-
-/**
- * Process a single stream-json event: update mutable state, write to log file,
- * and forward filtered summaries to terminal. Extracted from executeClaudeProcess()
- * to enable unit testing without spawning the Claude CLI.
- */
-export async function processStreamEvent(
-  // deno-lint-ignore no-explicit-any
-  event: Record<string, any>,
-  state: StreamProcessorState,
-): Promise<void> {
-  if (event.type === "assistant") {
-    state.turnCount++;
-    if (state.logFile) {
-      await state.logFile.write(
-        state.encoder.encode(
-          stampLines(`--- turn ${state.turnCount} ---`) + "\n",
-        ),
-      );
-    }
-    const contents = event.message?.content;
-    if (Array.isArray(contents)) {
-      for (const block of contents) {
-        if (block.type === "tool_use" && block.name === "Read") {
-          const warn = state.tracker.track(block.input?.file_path);
-          if (warn && state.logFile) {
-            await state.logFile.write(
-              state.encoder.encode(stampLines(warn) + "\n"),
-            );
-          }
-        }
-      }
-    }
-  }
-  if (event.type === "result") {
-    state.resultEvent = extractClaudeOutput(event);
-  }
-  const logSummary = formatEventForOutput(event);
-  if (state.logFile && logSummary) {
-    await state.logFile.write(
-      state.encoder.encode(stampLines(logSummary) + "\n"),
-    );
-  }
-  if (event.type === "result" && state.resultEvent && state.logFile) {
-    await state.logFile.write(
-      state.encoder.encode(stampLines("--- end ---") + "\n"),
-    );
-    await state.logFile.write(
-      state.encoder.encode(stampLines(formatFooter(state.resultEvent)) + "\n"),
-    );
-  }
-  if (state.onOutput) {
-    const termSummary = formatEventForOutput(event, state.verbosity);
-    if (termSummary) state.onOutput(termSummary);
-  }
-}
-
-/**
- * Execute the claude CLI process with stream-json output.
- * Processes NDJSON events in real-time: writes readable formatted summaries
- * to streamLogPath (full, unfiltered), forwards filtered summaries to onOutput
- * (tool_use suppressed when verbosity=semi-verbose), and extracts ClaudeCliOutput
- * from the final "result" event.
- */
-async function executeClaudeProcess(
-  args: string[],
-  timeoutSeconds: number,
-  onOutput?: (line: string) => void,
-  streamLogPath?: string,
-  verbosity?: Verbosity,
-): Promise<ClaudeCliOutput> {
-  // Unset CLAUDECODE to allow nested claude CLI invocations.
-  // Claude Code checks this variable and refuses to launch inside another session.
-  // Deno.Command merges env with parent, so setting empty string overrides it.
-  const cmd = new Deno.Command("claude", {
-    args,
-    stdin: "null",
-    stdout: "piped",
-    stderr: "piped",
-    env: { CLAUDECODE: "" },
-  });
-
-  const process = cmd.spawn();
-  register(process);
-
-  try {
-    // Set up timeout
-    const timeoutId = setTimeout(() => {
-      try {
-        process.kill("SIGTERM");
-      } catch {
-        // Process may have already exited
-      }
-    }, timeoutSeconds * 1000);
-
-    // Open stream log file for real-time writing (append mode)
-    let logFile: Deno.FsFile | undefined;
-    if (streamLogPath) {
-      const dir = streamLogPath.replace(/\/[^/]+$/, "");
-      await Deno.mkdir(dir, { recursive: true });
-      logFile = await Deno.open(streamLogPath, {
-        write: true,
-        create: true,
-        append: true,
-      });
-    }
-
-    // Process stdout as stream-json NDJSON
-    const state: StreamProcessorState = {
-      turnCount: 0,
-      resultEvent: undefined,
-      tracker: new FileReadTracker(),
-      logFile,
-      encoder: new TextEncoder(),
-      onOutput,
-      verbosity,
-    };
-    const stdoutDecoder = new TextDecoder();
-    let buffer = "";
-
-    const stdoutReader = process.stdout.getReader();
-    const stdoutDone = (async () => {
-      try {
-        while (true) {
-          const { done, value } = await stdoutReader.read();
-          if (done) break;
-          buffer += stdoutDecoder.decode(value, { stream: true });
-          const lines = buffer.split("\n");
-          buffer = lines.pop()!;
-          for (const line of lines) {
-            if (!line.trim()) continue;
-            try {
-              // deno-lint-ignore no-explicit-any
-              const event = JSON.parse(line) as Record<string, any>;
-              await processStreamEvent(event, state);
-            } catch {
-              // Skip malformed JSON lines
-            }
-          }
-        }
-        // Process remaining buffer
-        if (buffer.trim()) {
-          try {
-            // deno-lint-ignore no-explicit-any
-            const event = JSON.parse(buffer) as Record<string, any>;
-            await processStreamEvent(event, state);
-          } catch { /* skip */ }
-        }
-      } catch { /* stream closed */ }
-    })();
-
-    // Collect stderr for error reporting
-    const stderrChunks: Uint8Array[] = [];
-    const stderrReader = process.stderr.getReader();
-    const stderrDone = (async () => {
-      try {
-        while (true) {
-          const { done, value } = await stderrReader.read();
-          if (done) break;
-          stderrChunks.push(value);
-        }
-      } catch { /* stream closed */ }
-    })();
-
-    await Promise.all([stdoutDone, stderrDone]);
-    const status = await process.status;
-    clearTimeout(timeoutId);
-
-    // Close log file
-    if (logFile) {
-      logFile.close();
-    }
-
-    const concat = (chunks: Uint8Array[]) => {
-      const total = chunks.reduce((n, c) => n + c.length, 0);
-      const buf = new Uint8Array(total);
-      let offset = 0;
-      for (const c of chunks) {
-        buf.set(c, offset);
-        offset += c.length;
-      }
-      return buf;
-    };
-    const stderr = new TextDecoder().decode(concat(stderrChunks)).trim();
-
-    if (state.resultEvent) {
-      return state.resultEvent;
-    }
-
-    if (!status.success) {
-      throw new Error(
-        `Claude CLI exited with code ${status.code}${
-          stderr ? `: ${stderr}` : ""
-        }`,
-      );
-    }
-
-    throw new Error(
-      "Claude CLI stream-json output contained no result event",
-    );
-  } finally {
-    unregister(process);
-  }
-}
-
-/** Extract ClaudeCliOutput fields from a stream-json result event. */
-export function extractClaudeOutput(
-  // deno-lint-ignore no-explicit-any
-  event: Record<string, any>,
-): ClaudeCliOutput {
-  return {
-    result: event.result ?? "",
-    session_id: event.session_id ?? "",
-    total_cost_usd: event.total_cost_usd ?? 0,
-    duration_ms: event.duration_ms ?? 0,
-    duration_api_ms: event.duration_api_ms ?? 0,
-    num_turns: event.num_turns ?? 0,
-    is_error: event.is_error ?? event.subtype !== "success",
-    permission_denials: event.permission_denials,
-  };
-}
-
-/** Shorten an absolute path by stripping common workspace prefixes. */
-function shortenPath(p: string): string {
-  return p.replace(/^\/workspaces\/[^/]+\//, "").replace(
-    /^\/[^/]+\/[^/]+\/[^/]+\/[^/]+\/[^/]+\//,
-    "",
-  );
-}
-
-const MAX_CMD_LEN = 80;
-
-/** Extract a human-readable detail string from a tool_use input. */
-// deno-lint-ignore no-explicit-any
-function formatToolDetail(name: string, input?: Record<string, any>): string {
-  if (!input) return "";
-  switch (name) {
-    case "Read":
-    case "Write":
-    case "Edit":
-      return input.file_path ? shortenPath(input.file_path) : "";
-    case "Bash":
-      if (input.description) return input.description;
-      if (input.command) {
-        const cmd = input.command as string;
-        return cmd.length > MAX_CMD_LEN
-          ? `\`${cmd.slice(0, MAX_CMD_LEN)}…\``
-          : `\`${cmd}\``;
-      }
-      return "";
-    case "Grep":
-      return [
-        input.pattern ? `/${input.pattern}/` : "",
-        input.path ? `in ${shortenPath(input.path)}` : "",
-      ].filter(Boolean).join(" ");
-    case "Glob":
-      return input.pattern ?? "";
-    case "Agent":
-      return input.description ?? "";
-    default:
-      return "";
-  }
-}
-
-/**
- * Format a stream event as a one-line summary for output.
- * When verbosity is "semi-verbose", tool_use blocks in assistant events are
- * suppressed — only text blocks are emitted. Default undefined = all blocks.
- * Log file writes call without verbosity to preserve full output.
- */
-export function formatEventForOutput(
-  // deno-lint-ignore no-explicit-any
-  event: Record<string, any>,
-  verbosity?: Verbosity,
-): string {
-  switch (event.type) {
-    case "system":
-      if (event.subtype === "init") {
-        return `[stream] init model=${event.model ?? "?"}`;
-      }
-      return "";
-    case "assistant": {
-      const contents = event.message?.content;
-      if (!Array.isArray(contents)) return "";
-      const parts: string[] = [];
-      for (const block of contents) {
-        if (block.type === "text" && block.text) {
-          const preview = block.text.length > 120
-            ? block.text.slice(0, 120) + "…"
-            : block.text;
-          parts.push(`[stream] text: ${preview.replaceAll("\n", "↵")}`);
-        } else if (block.type === "tool_use") {
-          if (verbosity === "semi-verbose") continue;
-          const detail = formatToolDetail(block.name, block.input);
-          parts.push(
-            detail
-              ? `[stream] tool: ${block.name ?? "?"} ${detail}`
-              : `[stream] tool: ${block.name ?? "?"}`,
-          );
-        }
-      }
-      return parts.join("\n");
-    }
-    case "result":
-      return `[stream] result: ${event.subtype} (${
-        event.duration_ms ?? 0
-      }ms, $${(event.total_cost_usd ?? 0).toFixed(4)})`;
-    default:
-      return "";
-  }
-}
-
 /** Run a shell command (for before/after hooks). */
 async function runShellCommand(
   command: string,
@@ -758,40 +294,4 @@ function toVerboseValidation(
     passed: r.passed,
     detail: r.message,
   }));
-}
-
-/**
- * Format a one-line summary footer for a completed Claude CLI run.
- * Pure function — unit-testable without CLI.
- * Format: `status=<ok|error> duration=<X>s cost=$<Y> turns=<N>`
- */
-export function formatFooter(output: ClaudeCliOutput): string {
-  const status = output.is_error ? "error" : "ok";
-  const duration = (output.duration_ms / 1000).toFixed(1);
-  const cost = output.total_cost_usd.toFixed(4);
-  return `status=${status} duration=${duration}s cost=$${cost} turns=${output.num_turns}`;
-}
-
-function sleep(ms: number): Promise<void> {
-  return new Promise((resolve) => setTimeout(resolve, ms));
-}
-
-/** Returns current time as [HH:MM:SS] prefix string. */
-export function tsPrefix(): string {
-  const d = new Date();
-  const h = String(d.getHours()).padStart(2, "0");
-  const m = String(d.getMinutes()).padStart(2, "0");
-  const s = String(d.getSeconds()).padStart(2, "0");
-  return `[${h}:${m}:${s}]`;
-}
-
-/**
- * Prepend timestamp to each non-empty line of text.
- * Empty lines pass through unchanged.
- */
-export function stampLines(text: string): string {
-  return text
-    .split("\n")
-    .map((line) => line ? `${tsPrefix()} ${line}` : line)
-    .join("\n");
 }
